Added Antialiasing

+++ includes/ray.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ray.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cfeijoo <cfeijoo@student.42.fr>            +#+  +:+       +#+        */
+/*   By: kube <kube@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/03/07 16:30:00 by lbinet            #+#    #+#             */
-/*   Updated: 2014/03/23 22:59:51 by cfeijoo          ###   ########.fr       */
+/*   Updated: 2014/03/24 23:50:02 by kube             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct						s_ray
 	float							inter_t;
 }									t_ray;
 
-t_ray		get_ray_from_point(int i, int j);
+t_ray		get_ray_from_point(float i, float j);
 void		throw_ray(t_ray *ray, int calculate_light,
 						t_object *to_ignore, int recursivity);
 void		throw_ray_predefined(t_ray *ray, int calculate_light,
